% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfoil_cis.R
\name{allfoil_cihigh}
\alias{allfoil_cihigh}
\title{Confidence Intervals for Proportion}
\usage{
allfoil_cihigh(linetabprops, sumlineup)
}
\arguments{
\item{linetabprops}{A dataframe of bootstrapped lineup proportions}

\item{sumlineup}{Number of members in a lineup}
}
\description{
Function to compute ci high for each foil in a lineup
}
\examples{
# Upper CI bounds for the choice proportions of three lineup members,
# based on 20 mock witnesses
allfoil_cihigh(c(0.5, 0.3, 0.2), 20)
}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. \emph{Law and Human Behavior, 5}(4), 299-309.
}
