% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfoilbias.R
\name{allfoilbias}
\alias{allfoilbias}
\title{Bias for each lineup member}
\usage{
allfoilbias(lineup_table)
}
\arguments{
\item{lineup_table}{A table of lineup choices}
}
\description{
Function to compute bias for each lineup member (assuming foil is suspect,
from Malpass, 1981)
}
\examples{
lineup_vec <- round(runif(100, 1, 6))
lineup_table <- table(lineup_vec)
x <- allfoilbias(lineup_table)

}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. Law and Human Behavior, 5(4), 299-309.
}
