% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGTmock.R
\docType{data}
\name{CGTmock}
\alias{CGTmock}
\title{GGTmock}
\format{An object of class \code{data.frame} with 133 rows and 3 columns.}
\usage{
data(CGTmock)
}
\description{
A data file containing lineup data for 3 independent lineups.
             133 rows x 3 columns. Each lineup trial consisted of 133 mock witnesses
            & 8 lineup members.

            \itemize{
            \item: Target position for lineup 1 is 3
            \item: Target position for lineup 2 is 6
            \item: Target position for lineup 3 is 2
            }
}
\keyword{datasets}
