% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp_functions.R
\name{compare_dpp}
\alias{compare_dpp}
\title{Compare DPP Between Two Procedures}
\usage{
compare_dpp(data_proc_a, data_proc_b, lineup_size = 6)
}
\arguments{
\item{data_proc_a}{Dataframe for procedure A (standard lineup format)}

\item{data_proc_b}{Dataframe for procedure B (standard lineup format)}

\item{lineup_size}{Integer. Lineup size (default = 6)}
}
\value{
A list containing:
  \itemize{
    \item dpp_a: DPP for procedure A
    \item dpp_b: DPP for procedure B
    \item dpp_difference: DPP_A - DPP_B (negative = A is better)
    \item dpp_obj_a: Full DPP object for procedure A
    \item dpp_obj_b: Full DPP object for procedure B
  }
}
\description{
Computes and compares DPP for two lineup procedures.
}
\details{
Compares two procedures using DPP. Since lower DPP indicates better
performance, a negative difference (DPP_A - DPP_B < 0) means procedure A
is better.

DPP comparison is advantageous when:
\itemize{
  \item ROC curves have different truncation points
  \item Confidence distributions differ between procedures
  \item pAUC comparisons might be misleading
}
}
\examples{
data(lineup_example)
# Compare two halves of the example data (odd vs even rows)
odd <- seq(1, nrow(lineup_example), by = 2)
cmp <- compare_dpp(lineup_example[odd, ], lineup_example[-odd, ])
cmp$dpp_difference

}
