% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauc_comparison.R
\name{.bootstrap_pauc_comparison}
\alias{.bootstrap_pauc_comparison}
\title{Bootstrap pAUC Comparison}
\usage{
.bootstrap_pauc_comparison(
  data1,
  data2,
  lineup_size,
  max_false_id_rate,
  n_bootstrap
)
}
\arguments{
\item{data1}{First dataset}

\item{data2}{Second dataset}

\item{lineup_size}{Lineup size}

\item{max_false_id_rate}{Maximum false ID rate}

\item{n_bootstrap}{Number of bootstrap samples}
}
\value{
List with bootstrap distributions
}
\description{
Bootstrap pAUC Comparison
}
\keyword{internal}
