% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zroc_sdt.R
\name{.bootstrap_sdt_fit}
\alias{.bootstrap_sdt_fit}
\title{Bootstrap SDT Parameter Estimates}
\usage{
.bootstrap_sdt_fit(raw_data, lineup_size, model, n_bootstrap, conf_level)
}
\description{
Bootstrap SDT Parameter Estimates
}
\keyword{internal}
