% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht.R
\name{.compute_2ht_probs}
\alias{.compute_2ht_probs}
\title{Compute 2-HT Model Predicted Probabilities}
\usage{
.compute_2ht_probs(dP, dA, b, g, L)
}
\arguments{
\item{dP}{Detection of culprit presence}

\item{dA}{Detection of culprit absence}

\item{b}{Biased selection}

\item{g}{Guessing}

\item{L}{Lineup size}
}
\value{
List with tp (target-present) and ta (target-absent) probability vectors
}
\description{
Compute 2-HT Model Predicted Probabilities
}
\keyword{internal}
