% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht_boot.R
\name{.compute_bca_ci}
\alias{.compute_bca_ci}
\title{Compute BCa Confidence Intervals}
\usage{
.compute_bca_ci(boot_estimates, original_est, conf_level)
}
\arguments{
\item{boot_estimates}{Matrix of bootstrap estimates}

\item{original_est}{Vector of original parameter estimates}

\item{conf_level}{Confidence level}
}
\value{
Matrix of confidence intervals
}
\description{
Compute BCa Confidence Intervals
}
\keyword{internal}
