% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauc_comparison.R
\name{.compute_pauc_with_cutoff}
\alias{.compute_pauc_with_cutoff}
\title{Compute pAUC with False ID Rate Cutoff}
\usage{
.compute_pauc_with_cutoff(roc_data, max_false_id_rate)
}
\arguments{
\item{roc_data}{Dataframe with roc data (from make_rocdata)}

\item{max_false_id_rate}{Maximum false ID rate cutoff}
}
\value{
Numeric pAUC value
}
\description{
Compute pAUC with False ID Rate Cutoff
}
\keyword{internal}
