% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht.R
\name{.extract_counts_from_df}
\alias{.extract_counts_from_df}
\title{Extract Counts from Data Frame}
\usage{
.extract_counts_from_df(data, target_present, identification)
}
\arguments{
\item{data}{Data frame with lineup identification data}

\item{target_present}{Column name for target presence indicator}

\item{identification}{Column name for identification decision}
}
\value{
Named vector of counts
}
\description{
Extract Counts from Data Frame
}
\keyword{internal}
