% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zroc_sdt.R
\name{.extract_sdt_parameters}
\alias{.extract_sdt_parameters}
\title{Extract SDT Parameters from z-ROC Fit}
\usage{
.extract_sdt_parameters(fit_results, zroc_data, model)
}
\description{
Extract SDT Parameters from z-ROC Fit
}
\keyword{internal}
