% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht_boot.R
\name{.reconstruct_raw_data}
\alias{.reconstruct_raw_data}
\title{Reconstruct Raw Data from winter_2ht Object}
\usage{
.reconstruct_raw_data(x)
}
\arguments{
\item{x}{A winter_2ht object}
}
\value{
Data frame with individual observations
}
\description{
Reconstruct Raw Data from winter_2ht Object
}
\keyword{internal}
