% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_sim.R
\name{face_sim}
\alias{face_sim}
\title{Compute similarity of faces in a lineup; experimental function}
\usage{
face_sim()
}
\description{
Function to compute the degree to which each face in a set of faces
loads onto a common factor computed from the faces.
}
\details{
\itemize{
        \item The faces need to be standardized for inter-pupil distance, and for
               pupil location prior to running the function

        \item The user will be asked to choose a set of faces through a dialog
              box. These should be jpeg files.

       \item There is no argument to the function; it gets what it needs from the
             dialog box.  The function prints the lineup array to the viewer
             pane in R, and reports the loading of each face on the first common
             factor in a factor analysis (using fa in package psych)
}
}
\references{
Tredoux, C. (2002). A direct measure of facial similarity
           and its relation to human similarity perceptions.
           \emph{Journal of Experimental Psychology: Applied, 8}(3), 180-193.
           doi:10.1037/1076-898x.8.3.180
}
