% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_samples.R
\name{gen_boot_samples}
\alias{gen_boot_samples}
\title{Bootstrap resampling}
\usage{
gen_boot_samples(lineup_vec, bootno)
}
\arguments{
\item{lineup_vec}{A numeric vectors of lineup choices}

\item{bootno}{Number of bootstrap samples}
}
\value{
A dataframe of bootstrapped lineup data
}
\description{
Function for generating bootstrapped samples from 1 vector of lineup data
}
\examples{
Data:
lineup_vec <- round(runif(100,1,6))
bootno <- 1000

Call:
bootdf <- gen_boot_samples(lineup_vec, bootno)

Execute function line-by-line:
names_a <- rep("sample_",1000)
names_b <- as.character(1:1000)
1000 \%>\% rerun(sample(lineup_vec, length(lineup_vec), replace = TRUE))  -> x
names(x) <- names_a
names(x) <- paste(names_a,names_b,sep = "")
bootdf <- map_df(x, magrittr:extract, c(1:7))
}
