% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_esize_m_ci.R
\name{gen_esize_m_ci}
\alias{gen_esize_m_ci}
\title{Bootstrap Quantile for Effective Size}
\usage{
gen_esize_m_ci(lineupsizes, perc = 0.05)
}
\arguments{
\item{lineupsizes}{A non-empty numeric vector of bootstrapped effective sizes.}

\item{perc}{A single quantile probability between 0 and 1. Defaults to 0.05.}
}
\value{
A single named bootstrap quantile.
}
\description{
Computes one requested quantile from a vector of bootstrapped Malpass effective
sizes. Call the function twice (for example, at 0.025 and 0.975) to obtain both
endpoints of a percentile interval.
}
\examples{
#Data:
lineup_vec <- rep(1:6, length.out = 100)
k <- 6

#Use gen_boot_samples to get bootstrapped data:
bootdata <- gen_boot_samples(lineup_vec, 1000)

#Compute effective size over df of bootstrapped data:
lineupsizes <- gen_esize_m(bootdata, 6)

#Call:
gen_esize_m_ci(lineupsizes)
gen_esize_m_ci(lineupsizes, perc = .025)
gen_esize_m_ci(lineupsizes, perc = .975)

}
