% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lineup_prop.R
\name{gen_lineup_prop}
\alias{gen_lineup_prop}
\title{Lineup proportion over dataframe}
\usage{
gen_lineup_prop(lineup_boot_df, target_pos)
}
\arguments{
\item{lineup_boot_df}{Dataframe of lineup choices (usually a bootstrapped set)}

\item{target_pos}{A scalar, representing target position in lineup. Must be declared by user}
}
\description{
Function for computing lineup proportion over a dataframe
}
\examples{
lineup_boot_df <- gen_boot_samples(lineup_vec, 1000)
target_pos <- 3
gen_lineup_prop(lineup_boot_df, target_pos)
bootprops <- gen_lineup_prop(lineup_boot_df, 3)

}
