% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_to_vec.R
\name{gen_linevec}
\alias{gen_linevec}
\title{Lineup vector}
\usage{
gen_linevec(lineup_table, k)
}
\arguments{
\item{lineup_table}{A table of lineup choices}

\item{k}{Nominal size (i.e.,total number of lineup members).
Must be declared by user}
}
\value{
Returns a vector of lineup choices
}
\description{
A function that takes a table of lineup choices and generates a vector of raw
lineup data
}
\examples{
#Data:
lineup_vec <- round(runif(100, 1, 6))
lineup_table <- table(lineup_vec)

#Call:
lineup_vec <- gen_linevec(lineup_table, 3)
}
