% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_to_vec.R
\name{gen_linevec_table}
\alias{gen_linevec_table}
\title{Lineup vector}
\usage{
gen_linevec_table(lineup_table, nom_size)
}
\arguments{
\item{lineup_table}{A table of lineup choices}

\item{nom_size}{Nominal size (i.e.,total number of lineup members).
Must be declared by user}
}
\value{
Returns a vector of lineup choices
}
\description{
A function that takes a table of lineup choices and generates a vector of raw
lineup data
}
\examples{
lineup_vec <- gen_linevec_table(lineup_table, 3)
}
