% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_mediapipe}
\alias{get_mediapipe}
\title{Get mediapipe module (may be NULL if not installed)}
\usage{
get_mediapipe()
}
\description{
Get mediapipe module (may be NULL if not installed)
}
\keyword{internal}
