% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\name{has_valid_face}
\alias{has_valid_face}
\title{Check if Image Contains a Valid Face}
\usage{
has_valid_face(img_path, detector = "retinaface", min_confidence = 0.5)
}
\arguments{
\item{img_path}{Path to the image file.}

\item{detector}{Detector to use.}

\item{min_confidence}{Minimum detection confidence (0-1).}
}
\value{
Logical. TRUE if a valid face is detected.
}
\description{
Quick check to determine if an image contains a detectable face
that can be used for embedding extraction.
}
\examples{
\dontrun{
if (has_valid_face("image.jpg")) {
  emb <- get_embedding("image.jpg")
}
}

}
