% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line73.R
\docType{data}
\name{line73}
\alias{line73}
\title{line73}
\format{An object of class \code{data.frame} with 42 rows and 1 columns.}
\usage{
data(line73)
}
\description{
A dataframe of lineup choices for 1 trial; real data from a Doob & Kirshenbaum (1973) article.
            This article was the first to discuss lineup size
            1 column x 42 observations

            42 mock witnesses
            Nominal size is 6
            Target position is 4
}
\references{
Doob, A. N., & Kirshenbaum, H. M. (1973). Bias in police lineups-Partial remembering.
       \emph{Journal of Police Science and Administration, 1}, 287-293.
}
\keyword{datasets}
