% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_prop_tab.R
\name{lineup_prop_tab}
\alias{lineup_prop_tab}
\title{Lineup proportion}
\usage{
lineup_prop_tab(lineup_table, target_pos)
}
\arguments{
\item{lineup_table}{A table of lineup choices}

\item{target_pos}{A scalar, representing target position in lineup. Must be declared by user}
}
\description{
Computes the proportion of mock witnesses identifying a particular lineup member
}
\examples{
lineup_prop_vec(lineup_table, 3)
lineup_prop_vec(table(lineup_vec), 2))
lineup_prop_vec(table(c(1, 2, 5, 3, 2, 4)), 2)
}
