% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_functions.R
\name{make_roc_gg}
\alias{make_roc_gg}
\title{Plot ROC Curve for Lineup Identification}
\usage{
make_roc_gg(rocobj_list, show_pauc = TRUE, point_labels = TRUE)
}
\arguments{
\item{rocobj_list}{List output from make_rocdata()}

\item{show_pauc}{Logical. Whether to display pAUC value on plot (default = TRUE)}

\item{point_labels}{Logical. Whether to label points with confidence levels (default = TRUE)}
}
\value{
A ggplot2 object
}
\description{
Creates a ggplot2 visualization of ROC data from lineup identification experiments.
}
\examples{
data(lineup_example)
roc <- make_rocdata(lineup_example)
make_roc_gg(roc)

}
