% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mickwick.R
\docType{data}
\name{mickwick}
\alias{mickwick}
\title{Confidence & Accuracy data (Mickes & Wixted)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 2 columns.}
\usage{
data(mickwick)
}
\description{
Data capturing confidence and accuracy for a lineup trial
}
\details{
mickwick
}
\keyword{datasets}
