% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esize_T_bayes.R
\name{plot.esize_T_bayes}
\alias{plot.esize_T_bayes}
\title{Plot method for esize_T_bayes objects}
\usage{
\method{plot}{esize_T_bayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"esize_T_bayes"}.}

\item{...}{Further arguments (unused).}
}
\value{
A ggplot object.
}
\description{
Plots the posterior distribution of Tredoux's E' as a histogram overlaid
with a kernel density estimate. Vertical lines mark the posterior median
and credible interval bounds.
}
