% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esize_T_bayes.R
\name{plot.esize_T_bayes_compare}
\alias{plot.esize_T_bayes_compare}
\title{Plot method for esize_T_bayes_compare objects}
\usage{
\method{plot}{esize_T_bayes_compare}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"esize_T_bayes_compare"}.}

\item{...}{Further arguments (unused).}
}
\value{
A ggplot object.
}
\description{
Plots the posterior distribution of Delta = E'_A - E'_B as a histogram
with density overlay. A vertical line at zero and shaded probability mass
above/below zero are shown.
}
