% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lineup_data.R
\name{plot.power_analysis}
\alias{plot.power_analysis}
\title{Plot Power Analysis Results}
\usage{
\method{plot}{power_analysis}(x, ...)
}
\arguments{
\item{x}{A power_analysis object}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot Power Analysis Results
}
