% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zroc_sdt.R
\name{plot.sdt_roc_fit}
\alias{plot.sdt_roc_fit}
\title{Plot Method for SDT ROC Fit}
\usage{
\method{plot}{sdt_roc_fit}(x, ...)
}
\arguments{
\item{x}{An sdt_roc_fit object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot Method for SDT ROC Fit
}
