% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht_plots.R
\name{plot.winter_2ht}
\alias{plot.winter_2ht}
\title{Plot Method for winter_2ht Objects}
\usage{
\method{plot}{winter_2ht}(x, ...)
}
\arguments{
\item{x}{A winter_2ht object}

\item{...}{Additional arguments passed to plot_2ht_parameters()}
}
\description{
Creates a default plot for winter_2ht objects showing parameter estimates.
}
