% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht_boot.R
\name{plot.winter_2ht_boot}
\alias{plot.winter_2ht_boot}
\title{Plot Bootstrap Distributions}
\usage{
\method{plot}{winter_2ht_boot}(x, ...)
}
\arguments{
\item{x}{A winter_2ht_boot object}

\item{...}{Additional arguments (not used)}
}
\description{
Creates histograms showing the bootstrap distributions of parameter estimates.
}
