% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esize_T_bayes.R
\name{print.esize_T_bayes}
\alias{print.esize_T_bayes}
\title{Print method for esize_T_bayes objects}
\usage{
\method{print}{esize_T_bayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"esize_T_bayes"}.}

\item{...}{Further arguments (unused).}
}
\description{
Print method for esize_T_bayes objects
}
