% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anri_functions.R
\name{print.lineup_anri}
\alias{print.lineup_anri}
\title{Print Method for ANRI Objects}
\usage{
\method{print}{lineup_anri}(x, ...)
}
\arguments{
\item{x}{An anri object from compute_anri()}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for ANRI Objects
}
