% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anri_functions.R
\name{print.lineup_anri_comparison}
\alias{print.lineup_anri_comparison}
\title{Print Method for ANRI Comparison Objects}
\usage{
\method{print}{lineup_anri_comparison}(x, ...)
}
\arguments{
\item{x}{A comparison object from compare_anri()}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for ANRI Comparison Objects
}
