% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_curves_functions.R
\name{print.lineup_bayes_curves}
\alias{print.lineup_bayes_curves}
\title{Print Method for Bayes Curves Objects}
\usage{
\method{print}{lineup_bayes_curves}(x, ...)
}
\arguments{
\item{x}{A bayes_curves object from make_bayes_curves()}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Bayes Curves Objects
}
