% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cac_functions.R
\name{print.lineup_cac}
\alias{print.lineup_cac}
\title{Print Method for lineup_cac Objects}
\usage{
\method{print}{lineup_cac}(x, ...)
}
\arguments{
\item{x}{A lineup_cac object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_cac Objects
}
