% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{print.lineup_calibration}
\alias{print.lineup_calibration}
\title{Print Method for lineup_calibration Objects}
\usage{
\method{print}{lineup_calibration}(x, ...)
}
\arguments{
\item{x}{A lineup_calibration object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_calibration Objects
}
