% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{print.lineup_data}
\alias{print.lineup_data}
\title{Print Method for Lineup Data}
\usage{
\method{print}{lineup_data}(x, ...)
}
\arguments{
\item{x}{A lineup_data object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Lineup Data
}
