% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp_functions.R
\name{print.lineup_dpp}
\alias{print.lineup_dpp}
\title{Print Method for lineup_dpp Objects}
\usage{
\method{print}{lineup_dpp}(x, ...)
}
\arguments{
\item{x}{A lineup_dpp object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_dpp Objects
}
