% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig_functions.R
\name{print.lineup_eig}
\alias{print.lineup_eig}
\title{Print Method for lineup_eig Objects}
\usage{
\method{print}{lineup_eig}(x, n_responses = 10, ...)
}
\arguments{
\item{x}{A lineup_eig object from compute_eig()}

\item{n_responses}{Number of response categories to display (default = 10)}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_eig Objects
}
