% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullroc_functions.R
\name{print.lineup_fullroc}
\alias{print.lineup_fullroc}
\title{Print Method for lineup_fullroc Objects}
\usage{
\method{print}{lineup_fullroc}(x, show_diagnosticity = TRUE, ...)
}
\arguments{
\item{x}{A lineup_fullroc object}

\item{show_diagnosticity}{Logical. Whether to print diagnosticity table (default = TRUE)}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_fullroc Objects
}
