% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_range_functions.R
\name{print.lineup_ppv}
\alias{print.lineup_ppv}
\title{Print Method for lineup_ppv Objects}
\usage{
\method{print}{lineup_ppv}(x, ...)
}
\arguments{
\item{x}{A lineup_ppv object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_ppv Objects
}
