% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_range_functions.R
\name{print.lineup_ppv_range}
\alias{print.lineup_ppv_range}
\title{Print Method for lineup_ppv_range Objects}
\usage{
\method{print}{lineup_ppv_range}(x, ...)
}
\arguments{
\item{x}{A lineup_ppv_range object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_ppv_range Objects
}
