% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_functions.R
\name{print.lineup_rac}
\alias{print.lineup_rac}
\title{Print Method for lineup_rac Objects}
\usage{
\method{print}{lineup_rac}(x, ...)
}
\arguments{
\item{x}{A lineup_rac object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_rac Objects
}
