% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_roc.R
\name{print.lineup_roc}
\alias{print.lineup_roc}
\title{Print Method for lineup_roc Objects}
\usage{
\method{print}{lineup_roc}(x, ...)
}
\arguments{
\item{x}{A lineup_roc object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_roc Objects
}
