% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{print.lineup_similarity}
\alias{print.lineup_similarity}
\title{Print Method for lineup_similarity}
\usage{
\method{print}{lineup_similarity}(x, ...)
}
\arguments{
\item{x}{A lineup_similarity object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_similarity
}
