% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.lineup_utility_comparison}
\alias{print.lineup_utility_comparison}
\title{Print Method for lineup_utility_comparison Objects}
\usage{
\method{print}{lineup_utility_comparison}(x, ...)
}
\arguments{
\item{x}{A lineup_utility_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for lineup_utility_comparison Objects
}
