% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{print.model_comparison}
\alias{print.model_comparison}
\title{Print Method for model_comparison Objects}
\usage{
\method{print}{model_comparison}(x, ...)
}
\arguments{
\item{x}{A model_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for model_comparison Objects
}
