% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauc_comparison.R
\name{print.pauc_comparison}
\alias{print.pauc_comparison}
\title{Print Method for pauc_comparison Objects}
\usage{
\method{print}{pauc_comparison}(x, ...)
}
\arguments{
\item{x}{A pauc_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for pauc_comparison Objects
}
