% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zroc_sdt.R
\name{print.sdt_roc_fit}
\alias{print.sdt_roc_fit}
\title{Print Method for SDT ROC Fit}
\usage{
\method{print}{sdt_roc_fit}(x, ...)
}
\arguments{
\item{x}{An sdt_roc_fit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for SDT ROC Fit
}
