% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lineup_data.R
\name{print.simulated_lineup_data}
\alias{print.simulated_lineup_data}
\title{Print Method for Simulated Lineup Data}
\usage{
\method{print}{simulated_lineup_data}(x, ...)
}
\arguments{
\item{x}{A simulated_lineup_data object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Simulated Lineup Data
}
