% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht.R
\name{print.winter_2ht}
\alias{print.winter_2ht}
\title{Print Method for winter_2ht Objects}
\usage{
\method{print}{winter_2ht}(x, ...)
}
\arguments{
\item{x}{A winter_2ht object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for winter_2ht Objects
}
