% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht_boot.R
\name{print.winter_2ht_boot}
\alias{print.winter_2ht_boot}
\title{Print Method for winter_2ht_boot Objects}
\usage{
\method{print}{winter_2ht_boot}(x, ...)
}
\arguments{
\item{x}{A winter_2ht_boot object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for winter_2ht_boot Objects
}
