% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_index.R
\name{rep_index}
\alias{rep_index}
\title{Rep index}
\usage{
rep_index(index, num)
}
\arguments{
\item{index}{Number to be repeated}

\item{num}{Number of times to repeat index}
}
\description{
Helper function that repeats a number k times
}
\details{
Can be used to create vectors of lineup data
}
\examples{
rep_index(1, 10)
}
