% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_vector.R
\name{rot_vector}
\alias{rot_vector}
\title{Rotate vector}
\usage{
rot_vector(lineup_vec)
}
\arguments{
\item{lineup_vec}{A vector of lineup data}
}
\description{
Helper function to rotate vector
}
\examples{
# Rotate lineup positions by one place
rot_vector(1:6)
}
