% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{summary.lineup_similarity}
\alias{summary.lineup_similarity}
\title{Summary Method for lineup_similarity}
\usage{
\method{summary}{lineup_similarity}(object, ...)
}
\arguments{
\item{object}{A lineup_similarity object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for lineup_similarity
}
