% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{summary.model_comparison}
\alias{summary.model_comparison}
\title{Summary Method for model_comparison Objects}
\usage{
\method{summary}{model_comparison}(object, ...)
}
\arguments{
\item{object}{A model_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for model_comparison Objects
}
