% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauc_comparison.R
\name{summary.pauc_comparison}
\alias{summary.pauc_comparison}
\title{Summary Method for pauc_comparison Objects}
\usage{
\method{summary}{pauc_comparison}(object, ...)
}
\arguments{
\item{object}{A pauc_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for pauc_comparison Objects
}
