% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zroc_sdt.R
\name{summary.sdt_roc_fit}
\alias{summary.sdt_roc_fit}
\title{Summary Method for SDT ROC Fit}
\usage{
\method{summary}{sdt_roc_fit}(object, ...)
}
\arguments{
\item{object}{An sdt_roc_fit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for SDT ROC Fit
}
