% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht.R
\name{summary.winter_2ht}
\alias{summary.winter_2ht}
\title{Summary Method for winter_2ht Objects}
\usage{
\method{summary}{winter_2ht}(object, ...)
}
\arguments{
\item{object}{A winter_2ht object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for winter_2ht Objects
}
