% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht_boot.R
\name{summary.winter_2ht_boot}
\alias{summary.winter_2ht_boot}
\title{Summary Method for winter_2ht_boot Objects}
\usage{
\method{summary}{winter_2ht_boot}(object, ...)
}
\arguments{
\item{object}{A winter_2ht_boot object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for winter_2ht_boot Objects
}
