% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfoil_cis.R
\name{allfoil_cihigh}
\alias{allfoil_cihigh}
\title{Confidence Intervals for Proportion}
\usage{
allfoil_cihigh(linetabprops, sumlineup)
}
\arguments{
\item{sumlineup}{Number of members in a lineup}

\item{lineuptabprops}{A dataframe of bootstrapped lineup proportions}
}
\description{
Function to compute ci high for each foil in a lineup
}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. \emph{Law and Human Behavior, 5}(4), 299-309.
}
