% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfoilbias.R
\name{allfoilbias}
\alias{allfoilbias}
\title{Bias for each lineup member}
\usage{
allfoilbias(lineup_table, target_pos, k)
}
\arguments{
\item{lineup_table}{A table of lineup choices}

\item{target_pos}{A scalar, representing target position in lineup. Must be declared by user}

\item{k}{Nominal size (i.e., total number of members in lineup). Must be specified by user (scalar).}
}
\description{
Function to compute bias for each lineup member (assuming foil is suspect,
from Malpass, 1981)
}
\examples{
lineup_vec <- round(runif(100, 1, 6))
lineup_table <- table(lineup_vec)
x <- allfoilbias(lineup_table, 5, 6)

}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. \emph{Law and Human Behavior, 5}(4), 299-309.
}
