% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_diag.R
\name{chi_diag}
\alias{chi_diag}
\title{Chi-squared estimate of homogeneity of diagnosticity ratio}
\usage{
chi_diag(df)
}
\arguments{
\item{df}{A dataframe containing: ln(d), variance of ln(d), d weights}
}
\description{
Function for getting chi-squared value for homogeneity of diagnosticity ratios
}
\details{
To compute df, use the diag_param helper function.
}
\examples{
ratio <- ln_diag_ratio(linedf)
var <- var_lnd(linedf)
wi <- d_weights(linedf)
df <- cbind(ratio, var, wi)
chi_diag(df)

}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. \emph{Law and Human Behavior, 5}(4), 299-309.

           Malpass, R. S., Tredoux, C., & McQuiston-Surrett, D. (2007). Lineup
           construction and lineup fairness. In R. Lindsay, D. F. Ross, J. D. Read,
           & M. P. Toglia (Eds.), \emph{Handbook of Eyewitness Psychology, Vol. 2: Memory for
           people} (pp. 155-178). Mahwah, NJ: Lawrence Erlbaum Associates.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup fairness.
           \emph{Law and Human Behavior, 22}(2), 217-237.

           Tredoux, C. (1999). Statistical considerations when determining measures of
           lineup size and lineup bias. \emph{Applied Cognitive Psychology, 13}, S9-S26.

           Wells, G. L., Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
           empirically assessing the fairness of a lineup. \emph{Law and Human Behavior,
           3}(4), 285-293.
}
