% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_ES_functions.R
\name{compare_eff_sizes.boot}
\alias{compare_eff_sizes.boot}
\title{Comparing Effective Size: Base function for bootstrapping}
\usage{
compare_eff_sizes.boot(linedf, d)
}
\arguments{
\item{linedf}{A dataframe of lineup data. Must consist of 2 columns, each
containing data for 2 independent lineups}

\item{d}{Indices for bootstrap sample. Argument used by boot function to 
select samples for bootstrapping}
}
\description{
A base function for bootstrapping a dataframe of choices for 2 independent lineups
}
\details{
The approach here is to compute the effective size of each lineup
        separately, and to take the difference between them. This is then 
        bootstrapped, and if the bootstrap does not contain 0, we
         conclude the effective size estimates are different at p = alpha
}
\examples{
Data:
}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in eyewitness
                 identification lineups. Law and Human Behavior, 5(4), 299-309.

           Malpass, R. S., Tredoux, C., & McQuiston-Surrett, D. (2007).
                 Lineup construction and lineup fairness. In R. Lindsay, D. F. Ross, J. D. Read, & M. P. Toglia (Eds.), Handbook of Eyewitness Psychology, Vol. 2: Memory for people (pp. 155-178). Mahwah, NJ: Lawrence Erlbaum Associates.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup
                 fairness. Law and Human Behavior, 22(2), 217-237.

           Tredoux, C. (1999). Statistical considerations when determining measures
                 of lineup size and lineup bias. Applied Cognitive Psychology, 13, S9-S26.

           Wells, G. L., Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
                 empirically assessing the fairness of a lineup. Law and Human Behavior, 3(4), 285-293.
}
