% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{cosine_to_similarity}
\alias{cosine_to_similarity}
\title{Convert Cosine Distance to Similarity Score}
\usage{
cosine_to_similarity(distance)
}
\arguments{
\item{distance}{Numeric. Cosine distance value(s).}
}
\value{
Numeric similarity score(s) where 1 = identical, 0 = maximally different.
}
\description{
Converts cosine distance to a similarity score in [0, 1] range.
}
\examples{
# Distance of 0 -> similarity of 1
cosine_to_similarity(0)  # Returns 1

# Distance of 0.5 -> similarity of 0.5
cosine_to_similarity(0.5)  # Returns 0.5

}
