% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_weights.R
\name{d_weights}
\alias{d_weights}
\title{Diagnosticity ratio weights}
\usage{
d_weights(linedf)
}
\arguments{
\item{linedf}{A dataframe of parameters for computing diagnosticity ratio}
}
\value{
A dataframe with one column containing weights for each pair of
        lineups for which diagnosticity ratio is being calculated.
}
\description{
Function to compute weights of each diagnosticity ratio  for k lineup pairs
}
\details{
In order to obtain a pooled estimate of a set of diagnosticity ratios,
        we use a weight for each ratio that is equal to the inverse of its
        variance.

        \strong{To get linedf, use the diag_param helper function}

         \emph{diag_param} returns a dataframe containing the following:

         \itemize{
         \item \emph{n11}: Number of mock witnesses who identified the suspect in the target
              present condition

         \item \emph{n21}: Number of mock witnesses who did not identify the suspect in the
             target present condition

         \item \emph{n12}: Number of mock witnesses who identified the suspect in the target
             absent condition

         \item \emph{n13}: Number of mock witnesses who did not identify the suspect in the
             target absent condition
             }
}
\examples{
#Data:
linedf <- diag_param(lineup_pres_list, lineup_abs_list, pos_pres, abs_pres)

#Call:
wi <- d_weights(linedf)

}
\references{
Tredoux, C. G. (1998). Statistical inference on measures of lineup
           fairness. \emph{Law and Human Behavior, 22}(2), 217-237.
}
