% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck1.R
\name{datacheck1}
\alias{datacheck1}
\title{Helper function}
\usage{
datacheck1(lineup_vec, k)
}
\arguments{
\item{lineup_vec}{A numeric vector of lineup choices}

\item{k}{Number of members in lineup}
}
\description{
Checks that number of lineup choices contained in a vector is accurate
}
\details{
This function ensures that a non-selected lineup member is not accidentally
        omitted from the dataframe due to lack of selection by all mock witnesses.
        It functions as a check that the total number of lineup members is accurate.
}
