% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck2.R
\name{datacheck2}
\alias{datacheck2}
\title{Helper function}
\usage{
datacheck2(lineup_pres, lineup_abs, k1, k2)
}
\arguments{
\item{lineup_pres}{A numeric vector of lineup choices for a lineup in which
the target was present}

\item{lineup_abs}{A numeric vector of lineup choices for a lineup in which
the target was absent}

\item{k1}{Number of targets in TP lineup}

\item{k2}{Number of targets in TA lineup}
}
\description{
Checks that number of lineup choices contained in a vector is accurate
}
\details{
This function ensures that a non-selected lineup member is not accidentally
        omitted from the dataframe due to lack of selection by all mock witnesses.
        It functions as a check that the total number of lineup members is accurate
}
