% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_param.R
\name{diag_param}
\alias{diag_param}
\title{Parameters for diagnosticity ratio}
\usage{
diag_param(lineup_pres_list, lineup_abs_list, pos_pres, pos_abs)
}
\arguments{
\item{lineup_pres_list}{A list containing k vectors of lineup choices for k lineups, in which the
target was present}

\item{lineup_abs_list}{A list containing k vectors of lineup choices for k lineups, in which the
target was absent}

\item{pos_pres}{A list containing k numeric vectors indexing lineup member positions 
for the target present condition}

\item{pos_abs}{A list containing k numeric vectors indexing lineup member positions 
for the target absent condition}
}
\description{
This function calculates the parameters needed to calculate the diagnosticity
        ratio for several lineup pairs.
}
\details{
Lineup pairs consist of one lineup in which the target was present (TP)
        and one lineup in which the target was absent (TA).

        Each lineup pair must occupy corresponding positions in the TA and TP lists.

        Example:

        For a lineup pair A that consists of (1)TP lineup and (2)TA lineup:
        A(1) is the first vector in the TP list
        A(2) is the first vector in the TP list
        
        Data must be in a list format. This allows the function to compare 
        lineups in which the number of choices and number of lineup members differs.
}
\examples{
#Target present data:
A <-  round(runif(100,1,6))
B <-  round(runif(70,1,5))
C <-  round(runif(20,1,4)) 
lineup_pos_list <- list(A, B, C)
rm(A, B, C)

a1 <- c(1, 2, 3, 4, 5, 6)
b1 <- c(1, 2, 3, 4, 5)
c1 <- c(1, 2, 3, 4)
pos_pres <- list(a1, b1, c1)
rm(a1, b1, c1)

Target absent data:
A <-  round(runif(100,1,6))
B <-  round(runif(70,1,5))
C <-  round(runif(20,1,4)) 
lineup_abs_list <- list(A, B, C)
rm(A, B, C)

a1 <- c(1, 2, 3, 4, 5, 6)
b1 <- c(1, 2, 3, 4, 5)
c1 <- c(1, 2, 3, 4)
pos_abs <- list(a1, b1, c1)
rm(a1, b1, c1)

Call:
diag_param(lineup_pres_list, lineup_abs_list, pos_pres, abs_pres)
}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. Law and Human Behavior, 5(4), 299-309.

           Malpass, R. S., Tredoux, C., & McQuiston-Surrett, D. (2007). Lineup
           construction and lineup fairness. In R. Lindsay, D. F. Ross, J. D. Read,
           & M. P. Toglia (Eds.), Handbook of Eyewitness Psychology, Vol. 2: Memory for
           people (pp. 155-178). Mahwah, NJ: Lawrence Erlbaum Associates.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup fairness.
           Law and Human Behavior, 22(2), 217-237.

           Tredoux, C. (1999). Statistical considerations when determining measures of
           lineup size and lineup bias. Applied Cognitive Psychology, 13, S9-S26.

           Wells, G. L.,Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
           empirically assessing the fairness of a lineup. Law and Human Behavior,
           3(4), 285-293.
}
