% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_size_per_foils.R
\name{eff_size_per_foils}
\alias{eff_size_per_foils}
\title{Effective Size per Foils}
\usage{
eff_size_per_foils(lineup_vec, target_pos)
}
\arguments{
\item{lineup_vec}{A numeric vector of lineup choices}

\item{target_pos}{A numeric vector indexing all lineup members}
}
\description{
Function for computing effective size (Malpass, 1981) counting foils who fall
within the CI for chance guessing
}
\examples{
#Data:
lineup_vec <- round(runif(100, 1, 6))
target_pos <- c(1, 2, 3, 4, 5, 6)

#Call:
eff_size_per_foils(lineup_vec, target_pos)
}
\references{
Malpass, R S. (1981). Effective size and defendant bias in eyewitness
           identification lineups. \emph{Law and Human Behavior, 5}, 299-309.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup
           fairness. \emph{Law and Human Behavior, 22}(2), 217-237.
}
