% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_ES_functions.R
\name{effsize_compare}
\alias{effsize_compare}
\title{Master Function: Comparing Effective Size}
\usage{
effsize_compare(linedf)
}
\arguments{
\item{linedf}{A dataframe of lineup data. Contains 2 columns, each of which
hold data for 2 independent lineups}
}
\description{
Function for comparing effective size of two indep lineups (Tredoux, 1998)
}
\details{
This function is a master function, calling other functions
        it needs, and reporting results in some detail

Function to call when bootstrap resampling using boot function (in package 'boot')
}
\references{
Davison,  A.C. & Hinkley,  D.V. (1997). Bootstrap methods and their
                 application. Cambridge University Press.

           Malpass, R. S. (1981). Effective size and defendant bias in eyewitness
                 identification lineups. Law and Human Behavior, 5(4), 299-309.

           Malpass, R. S., Tredoux, C., & McQuiston-Surrett, D. (2007).
                 Lineup construction and lineup fairness. In R. Lindsay, D. F. Ross, J. D. Read, & M. P. Toglia (Eds.), Handbook of Eyewitness Psychology, Vol. 2: Memory for people (pp. 155-178). Mahwah, NJ: Lawrence Erlbaum Associates.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup
                 fairness. Law and Human Behavior, 22(2), 217-237.

           Tredoux, C. (1999). Statistical considerations when determining measures
                 of lineup size and lineup bias. Applied Cognitive Psychology, 13, S9-S26.

           Wells, G. L., Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
                 empirically assessing the fairness of a lineup. Law and Human Behavior, 3(4), 285-293.
}
\seealso{
\code{\link[boot:boot]{boot}}: https://cran.r-project.org/web/packages/boot/boot.pdf
}
