% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esize_boot.R
\name{esize_m_boot}
\alias{esize_m_boot}
\title{Bootstrapped Effective Size}
\usage{
esize_m_boot(lineup_vec, d, printarg = FALSE)
}
\arguments{
\item{lineup_vec}{A vector of lineup choices}

\item{d}{Indices for bootstrap resampling}

\item{printarg}{Defaults to FALSE. If TRUE, provides both Tredoux's (1998)
and Malpass's (1981) calculations of effective size}
}
\description{
Base function for computing bootstrapped effective size
}
\details{
Function to call when bootstrap resampling using boot function (in package 'boot')
}
\references{
Davison,  A.C. & Hinkley,  D.V. (1997). Bootstrap methods and their
           application. Cambridge University Press.
           
           Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. Law and Human Behavior, 5(4), 299-309.

           Malpass, R. S., Tredoux, C., & McQuiston-Surrett, D. (2007). Lineup
           construction and lineup fairness. In R. Lindsay, D. F. Ross, J. D. Read,
           & M. P. Toglia (Eds.), Handbook of Eyewitness Psychology, Vol. 2: Memory for
           people (pp. 155-178). Mahwah, NJ: Lawrence Erlbaum Associates.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup fairness.
           Law and Human Behavior, 22(2), 217-237.

           Tredoux, C. (1999). Statistical considerations when determining measures of
           lineup size and lineup bias. Applied Cognitive Psychology, 13, S9-S26.

           Wells, G. L.,Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
           empirically assessing the fairness of a lineup. Law and Human Behavior,
           3(4), 285-293.
}
\seealso{
\code{\link[boot:boot]{boot}}: https://cran.r-project.org/web/packages/boot/boot.pdf
}
