% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_propmean_se.R
\name{gen_boot_propmean_se}
\alias{gen_boot_propmean_se}
\title{Descriptive statistics for bootstrapped lineup proportion}
\usage{
gen_boot_propmean_se(lineuprops)
}
\arguments{
\item{lineuprops}{A dataframe of bootstrapped lineup proportions}
}
\description{
Function for computing mean. med and se of boot proportion
}
\examples{

Data: Get bootstrapped data using gen_boot_samples?

}
