% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_samples.R
\name{gen_boot_samples}
\alias{gen_boot_samples}
\title{Bootstrap resampling}
\usage{
gen_boot_samples(lineup_vec, bootno)
}
\arguments{
\item{lineup_vec}{A numeric vectors of lineup choices}

\item{bootno}{Number of bootstrap samples}
}
\value{
A dataframe of bootstrapped lineup data
}
\description{
Function for generating bootstrapped samples from 1 vector of lineup data
}
\examples{
#Data:
lineup_vec <- round(runif(100,1,6))
bootno <- 1000

#Call:
bootdf <- gen_boot_samples(lineup_vec, bootno)

}
