% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_samples_list.R
\name{gen_boot_samples_list}
\alias{gen_boot_samples_list}
\title{Bootstrapped sampling}
\usage{
gen_boot_samples_list(lineup_list, bootno)
}
\arguments{
\item{lineup_list}{A list containing k vectors of lineup choices for k lineups,
in which the target was either absent or present}

\item{bootno}{Number of bootstrap samples}
}
\description{
Function for generating bootstrapped samples for of k vectors of lineup choices
}
