% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_esize_m_ci.R
\name{gen_esize_m_ci}
\alias{gen_esize_m_ci}
\title{Bootstrapped Confidence Intervals for Effective Size}
\usage{
gen_esize_m_ci(lineupesizes, perc = 0.05)
}
\arguments{
\item{perc}{Defaults to .05. Can be specified by user, according to desired
level of alpha (scalar)}

\item{lineupsizes}{A vector of bootstrapped effective sizes}
}
\value{
Confidence intervals for effective size
}
\description{
Function for computing Effective Size (Malpass, 1981, as adjusted by Tredoux, 1998)
with CIs from bootstrap df of lineups
}
\examples{
#Data:
esize_vec <- round(runif(100, 1, 6))

#Use gen_boot_samples to get bootstrapped data:
e_size_boot <- gen_boot_samples(esize_vec, 1000)

#Compute effective size for bootstrapped data:
lineupsizes <- lapply(bootdata, esize_m) \%>\% unlist

#Call:
gen_esize_m_ci(lineupsizes)
gen_esize_m_ci(lineupsizes, perc = .025)
gen_esize_m_ci(lineupsizes, perc = .975)
}
