% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_esize_m_ci.R
\name{gen_esize_m_ci}
\alias{gen_esize_m_ci}
\title{Bootstrapped Confidence Intervals for Effective Size}
\usage{
gen_esize_m_ci(lineupesizes, perc = 0.05)
}
\arguments{
\item{perc}{Defaults to 0.5. Can be specified by user, according to desired
level of alpha}

\item{lineupsizes}{A dataframe of bootstrapped effective sizes}
}
\value{
Confidence intervals for effective size
}
\description{
Function for computing Effective Size (Malpass, 1981, as adjusted by Tredoux, 1998)
with CIs from bootstrap df of lineups
}
\examples{
Use gen_boot_samples to get bootstrapped data

e_size_boot <- gen_boot_samples(esize_vec, 1000)
}
