% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\name{get_face_landmarks}
\alias{get_face_landmarks}
\title{Get 468-Point Face Mesh Landmarks}
\usage{
get_face_landmarks(img_path, return_3d = TRUE)
}
\arguments{
\item{img_path}{Path to the image file.}

\item{return_3d}{Logical. If TRUE, returns 3D coordinates (x, y, z).
If FALSE, returns only 2D coordinates (x, y). Default is TRUE.}
}
\value{
A list containing:
\itemize{
  \item landmarks - Matrix of landmark coordinates (468 x 2 or 468 x 3)
  \item face_detected - Logical indicating if a face was found
  \item image_width - Width of the input image
  \item image_height - Height of the input image
}
}
\description{
Extracts dense facial landmarks using MediaPipe Face Mesh,
which provides 468 3D landmark points covering the entire face.
}
\details{
MediaPipe Face Mesh provides detailed landmarks covering:
\itemize{
  \item Face oval/contour
  \item Left and right eyebrows
  \item Left and right eyes (including iris)
  \item Nose
  \item Lips (inner and outer contours)
  \item Face interior regions (cheeks, forehead)
}

The x and y coordinates are normalized to [0, 1] relative to image
dimensions. The z coordinate represents depth relative to the face center.
}
\note{
Requires MediaPipe to be installed (installed by default with
  install_faceR_python()).
}
\examples{
\dontrun{
# Get face mesh landmarks
mesh <- get_face_landmarks("face.jpg")

# Plot the landmarks
plot(mesh$landmarks[, 1], mesh$landmarks[, 2], pch = ".")
}

}
