% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homog_diag.R
\name{homog_diag}
\alias{homog_diag}
\title{Master function: Homogeneity of diagnosticity ratio}
\usage{
homog_diag(lineup_pres_list, lineup_abs_list, pos_pres, pos_abs)
}
\arguments{
\item{lineup_pres_list}{A list containing k vectors of lineup choices for k lineups, in which the
target was present}

\item{lineup_abs_list}{A list containing k vectors of lineup choices for k lineups, in which the
target was absent}

\item{pos_pres}{A numeric vector indexing lineup member positions for the target
present condition}

\item{pos_abs}{A numeric vector indexing lineup member positions for the target
absent condition}
}
\description{
This function provides assesses the homogeneity of the diagnosticity ratio of
 k lineup pairs
}
\details{
Computes diagnosticity ratio with chi-squared estimate and significance
        level for k lineup pairs
}
\examples{
#Target present data:
A <-  round(runif(100,1,6))
B <-  round(runif(70,1,5))
C <-  round(runif(20,1,4))
lineup_pres_list <- list(A, B, C)
rm(A, B, C)

a1 <- c(1, 2, 3, 4, 5, 6)
b1 <- c(1, 2, 3, 4, 5)
c1 <- c(1, 2, 3, 4)
pos_pres <- list(a1, b1, c1)
rm(a1, b1, c1)

Target absent data:
A <-  round(runif(100,1,6))
B <-  round(runif(70,1,5))
C <-  round(runif(20,1,4))
lineup_abs_list <- list(A, B, C)
rm(A, B, C)

a1 <- c(1, 2, 3, 4, 5, 6)
b1 <- c(1, 2, 3, 4, 5)
c1 <- c(1, 2, 3, 4)
pos_abs <- list(a1, b1, c1)
rm(a1, b1, c1)

Call:
homog_diag(lineup_pres_list, lineup_abs_list, pos_pres, abs_pres)
}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. Law and Human Behavior, 5(4), 299-309.

           Malpass, R. S., Tredoux, C., & McQuiston-Surrett, D. (2007). Lineup
           construction and lineup fairness. In R. Lindsay, D. F. Ross, J. D. Read,
           & M. P. Toglia (Eds.), Handbook of Eyewitness Psychology, Vol. 2: Memory for
           people (pp. 155-178). Mahwah, NJ: Lawrence Erlbaum Associates.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup fairness.
           Law and Human Behavior, 22(2), 217-237.

           Tredoux, C. (1999). Statistical considerations when determining measures of
           lineup size and lineup bias. Applied Cognitive Psychology, 13, S9-S26.

           Wells, G. L.,Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
           empirically assessing the fairness of a lineup. Law and Human Behavior,
           3(4), 285-293.
}
