% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_boot_allprop.R
\name{lineup_boot_allprop}
\alias{lineup_boot_allprop}
\title{Confidence intervals for lineup proportion}
\usage{
lineup_boot_allprop(lineup_vec, target_pos)
}
\arguments{
\item{lineup_vec}{A numeric vector of lineup choices}

\item{target_pos}{A numeric vector indexing all lineup members}
}
\value{
Returns a vector of bias corrected confidence intervals for 
       lineup proportion for each member in a lineup
}
\description{
Computes bootstrapped confidence intervals for lineup proportion
}
\details{
Function that computes bootstrapped lineup proportion using 1000 bootstrap draws
        Calls 'boot function in 'boot' package
}
\examples{
lineup_vec <- round(runif(100, 1, 6))
target_pos <- c(1, 2, 3, 4, 5, 6)
lineup_boot_allprop(lineup_vec, target_pos)
}
\references{
Davison,  A.C. & Hinkley,  D.V. (1997). Bootstrap methods and their
           application. Cambridge University Press.
           
           Wells, G. L., Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
           assessing the fairness of a lineup. Law and Human Behavior, 3(4),
           285-293.
}
\seealso{
\code{\link[boot:boot]{boot}}: https://cran.r-project.org/web/packages/boot/boot.pdf
}
