% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup_prop_boot.R
\name{lineup_prop_boot}
\alias{lineup_prop_boot}
\title{Bootstrapped lineup proportion}
\usage{
lineup_prop_boot(lineup_vec, d, target_pos)
}
\arguments{
\item{lineup_vec}{A numeric vector of lineup choices}

\item{target_pos}{A scalar, representing target position in lineup. Must be declared by user}

\item{d=d}{Indices for bootstrap sample. Argument used by boot function to 
select samples for bootstrapping}
}
\description{
Base function for computing bootstrapped lineup proportion for a lineup member
}
\details{
Function to call when bootstrap resampling using boot function
}
\references{
Davison,  A.C. & Hinkley,  D.V. (1997). Bootstrap methods and their
           application. Cambridge University Press
}
\seealso{
\code{\link[boot:boot]{boot}}: https://cran.r-project.org/web/packages/boot/boot.pdf
}
