% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_functions.R
\name{make_rocdata}
\alias{make_rocdata}
\title{Helper functions: Compute and plot ROC curve for lineup accuracy ~ confidence}
\usage{
make_rocdata(df_confacc)
}
\arguments{
\item{df_confacc}{A df with two columns, confidence and accuracy,
where accuracy = binary accuracy.}
}
\value{
An ROC object of package pROC
}
\description{
Function to compute and plot an ROC curve for data from an eyewitness
experiment, where accuracy is recorded for target present and target
absent lineups
}
