% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_helper.R
\name{makevec_prop}
\alias{makevec_prop}
\title{Helper functions}
\usage{
makevec_prop(prop, n)
}
\arguments{
\item{prop}{A dataframe of bootstrapped proportions}

\item{n}{Number of lineup members}
}
\description{
Several helper functions that compute bootcis from proportions
}
\examples{
# Expand a choice proportion of 0.25 among 20 mock witnesses into a
# binary choice vector
makevec_prop(0.25, 20)
}
