% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anri_functions.R
\name{plot_anri_difference_distribution}
\alias{plot_anri_difference_distribution}
\title{Plot Bootstrap Distribution of ANRI Difference}
\usage{
plot_anri_difference_distribution(compare_obj)
}
\arguments{
\item{compare_obj}{List output from compare_anri()}
}
\value{
A ggplot2 object
}
\description{
Creates a histogram of the bootstrap distribution of the difference
between two groups.
}
