% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_2ht_globals.R
\docType{package}
\name{r4lineups-package}
\alias{r4lineups}
\alias{r4lineups-package}
\title{r4lineups: Statistical Inference on Lineup Fairness}
\description{
Computes various properties of laboratory or police lineups for researchers in forensic psychology. The package includes functions for calculating lineup bias, functional size, effective size, diagnosticity ratio, homogeneity of the diagnosticity ratio, and ROC curves for confidence-accuracy data. Face similarity is computed using state-of-the-art deep learning embeddings (ArcFace, FaceNet, etc.) via Python's deepface library.
}
\author{
\strong{Maintainer}: Colin Tredoux \email{colin.tredoux@uct.ac.za}

Authors:
\itemize{
  \item Tamsyn Naylor \email{tmnaylor@outlook.com}
}

}
\keyword{internal}
