% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_lnd.R
\name{var_lnd}
\alias{var_lnd}
\title{Variance of ln of diagnosticity ratio}
\usage{
var_lnd(linedf)
}
\arguments{
\item{linedf}{A dataframe of parameters for computing diagnosticity ratio}
}
\description{
Function to compute variance of ln(d) for k lineup pairs
}
\details{
To get linedf, use the diag_param helper function

         diag_param: returns a dataframe containing the following:

         n11: Number of mock witnesses who identified the suspect in the target
              present condition

         n21: Number of mock witnesses who did not identify the suspect in the
             target present condition

         n12: Number of mock witnesses who identified the suspect in the target
             absent condition

         n13: Number of mock witnesses who did not identify the suspect in the
             target absent condition

Use diag_param helper function to get linedf: produces a dataframe containing
        the parameters needed to calculate diagnosticity ratio (see documentation for
        diag_param)

Calls var_d function
}
\examples{
linedf <- diag_param(lineup_pres_list, lineup_abs_list,
                                         pos_pres, abs_pres)
var_lnd <- var_lnd(linedf)
}
\references{
Malpass, R. S. (1981). Effective size and defendant bias in
           eyewitness identification lineups. Law and Human Behavior, 5(4), 299-309.

           Malpass, R. S., Tredoux, C., & McQuiston-Surrett, D. (2007). Lineup
           construction and lineup fairness. In R. Lindsay, D. F. Ross, J. D. Read,
           & M. P. Toglia (Eds.), Handbook of Eyewitness Psychology, Vol. 2: Memory for
           people (pp. 155-178). Mahwah, NJ: Lawrence Erlbaum Associates.

           Tredoux, C. G. (1998). Statistical inference on measures of lineup fairness.
           Law and Human Behavior, 22(2), 217-237.

           Tredoux, C. (1999). Statistical considerations when determining measures of
           lineup size and lineup bias. Applied Cognitive Psychology, 13, S9-S26.

           Wells, G. L.,Leippe, M. R., & Ostrom, T. M. (1979). Guidelines for
           empirically assessing the fairness of a lineup. Law and Human Behavior,
           3(4), 285-293.
}
